+++ application/classes/Core/ConfigurationException.class.php
@@ -6,7 +6,7 @@ namespace Core {
 	* @author Allan Thue Rehhoff
 	* @package Rehhoff_Framework
 	*/
-	class ConfigurationParserException extends \Exception {
+	class ConfigurationException extends \Exception {
 		public function __construct($message, $code = 0, Exception $previous = null) {
 			parent::__construct($message, $code, $previous);
 		}