+++ application/classes/Core/Controller.class.php
@@ -5,17 +5,41 @@ namespace Core {
 	class Controller {
 		protected $application, $request, $title;
 		protected $data = [];
+		private $theme;
 
 		public function __construct() {
 			$this->application = \Registry::get("Core\Application");
 			$this->request = array_merge($_GET, $_POST);
+			$this->setTitle("Frontpage");
+
+			$this->theme = (new Configuration($this->application->getThemepath()."/theme.json"));
+			$this->addThemeAssets();
 
 			$this->database = new \Database\Connection(
-				\Registry::get("Core\ConfigurationParser")->get("database.host"),
-				\Registry::get("Core\ConfigurationParser")->get("database.username"),
-				\Registry::get("Core\ConfigurationParser")->get("database.password"),
-				\Registry::get("Core\ConfigurationParser")->get("database.name")
+				\Registry::get("Core\Configuration")->get("database.host"),
+				\Registry::get("Core\Configuration")->get("database.username"),
+				\Registry::get("Core\Configuration")->get("database.password"),
+				\Registry::get("Core\Configuration")->get("database.name")
 			);
+		}
+
+		/**
+		* Adds configured theme assets to the DOM\Document class
+		* @uses \DOM\Document
+		* @return void
+		*/
+		private function addThemeAssets() {
+			if($this->theme->get("javascript")) {
+				foreach($this->theme->get("javascript") as $javascript) {
+					\DOM\Document::addJavascript(\Functions::url($javascript));
+				}
+			}
+
+			if($this->theme->get("stylesheets")) {
+				foreach($this->theme->get("stylesheets") as $stylesheet) {
+					\DOM\Document::addStylesheet(\Functions::url($stylesheet));
+				}
+			}
 		}
 
 		/**
@@ -32,7 +56,7 @@ namespace Core {
 		* @return self
 		*/
 		public function setTitle($title) {
-			$this->title = $title;
+			$this->data["title"] = sprintf(\Registry::get("Core\Configuration")->get("base_title"), $title);
 		}
 
 		/**
@@ -40,13 +64,7 @@ namespace Core {
 		* @return string
 		*/
 		public function getTitle() {
-			if(trim($this->title) != '') {
-				$title = sprintf($this->application->config->get("base_title"), $this->title);
-			} else {
-				$title = sprintf($this->application->config->get("base_title"), '');
-			}
-
-			return $title;
+			return $this->data["title"];
 		}
 	}
 }