+++ application/classes/Functions.class.php
@@ -2,7 +2,7 @@
 /**
 * @package Rehhoff_Framework
 */
-	class Debug {
+	class Functions {
 		/**
 		* Helps you debug your application.
 		* @author Allan Thue Rehhoff
@@ -16,13 +16,13 @@
 		* @param (boolean) $exit whether to exit the PHP process after printing debug information.
 		*/
 		public static function pre($stuff, $exit = false) {
-			echo "<pre style='background:#FFD; clear:both;'>";
+			print "<pre style='background:#FFD; clear:both;'>";
 			if(count($stuff) > 0 && $stuff !== false) {
 				print_r($stuff);
 			} else {
 				var_dump($stuff);
 			}
-			echo "</pre>";
+			print "</pre>";
 			
 			if($exit) exit;
 		}
@@ -46,6 +46,7 @@
 				$file = $trace[1]["file"];
 				$func = '';
 			}
+
 			if (isset($trace[3]["class"])) {
 				$class = $trace[3]["class"];
 				$func = $trace[3]["function"];
@@ -61,5 +62,20 @@
 			$c .= ($func != '') ? $func . "(): " : '';
 			
 			self::pre($c);
+		}
+
+		/**
+		* Removes any need for having a hardcoded basepath in some obscure place
+		* "cough"wordpress"cough"
+		* @param (string) $url Path to element of which to create a URI.
+		* @author Allan Thue Rehhoff
+		* @return string
+		*/
+		public static function url($asset = '') {
+			$protocol = SSL ? "https://" : "http://";
+			$host  = $_SERVER['HTTP_HOST'];
+			$path = rtrim(dirname($_SERVER['PHP_SELF']), '/\\');
+			$baseurl = $protocol.$host.$path."/";
+			return $baseurl.ltrim($asset, '/');
 		}
 	}