+++ application/controllers/Index.php
@@ -1,10 +1,10 @@
 <?php
-	class indexController extends \Core\Controller {
+	class IndexController extends \Core\Controller {
 		public function __construct() {
 			parent::__construct();
 		}
 
 		public function index() {
-			$this->data["intro"] = "html goes here.";
+			$this->data["intro"] = "A sample variable.";
 		}
 	}