+++ application/themes/default/header.tpl.php
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html>
 	<head>
-		<title><?php print $app->getTitle(); ?></title>
+		<title><?php print $title ?></title>
 		<?php foreach(\DOM\Document::getStylesheets() as $style): ?>
 			<link rel="stylesheet" type="text/css" href="<?php print $style; ?>" media="all" />
 		<?php endforeach; ?>